+++ CHANGELOG.md
@@ -7,6 +7,7 @@ version 0.10 - 20208-08-06
 - transport.file: targetFile is full path to file
 - isDebugActive
 - options.debug to active the debugger
+- level are limited to error, warn, info, debug'
 
 version 0.9
 - auto change transports to [transports]